smoketest scripts fixes (previous commit not completed) #2365

+++ tests/smoketest/scripts/run_replay.sh
@@ -13,6 +13,8 @@ then
 fi
 
 STEEMD_CONFIG=config.ini
+STEEMD_CONFIG_REF=config_ref.ini
+STEEMD_CONFIG_TEST=config_test.ini
 TEST_STEEMD_PATH=$1
 REF_STEEMD_PATH=$2
 TEST_WORK_PATH=$3
@@ -20,15 +22,53 @@ REF_WORK_PATH=$4
 BLOCK_LIMIT=$5
 RET_VAL1=-1
 RET_VAL2=-1
+EXIT_CODE=0
 
 function copy_config {
-   echo Copying ./$STEEMD_CONFIG over $1/$STEEMD_CONFIG
-   cp ./$STEEMD_CONFIG $1/$STEEMD_CONFIG
-   [ $? -ne 0 ] && echo FATAL: Failed to copy ./$STEEMD_CONFIG over $1/$STEEMD_CONFIG file. && exit -1
+   if [ -e $STEEMD_CONFIG ]
+   then
+      echo Copying ./$STEEMD_CONFIG over $1/$STEEMD_CONFIG
+      cp ./$STEEMD_CONFIG $1/$STEEMD_CONFIG
+      [ $? -ne 0 ] && echo FATAL: Failed to copy ./$STEEMD_CONFIG over $1/$STEEMD_CONFIG file. && exit -1
+   elif [ -e $2 ]
+   then
+      echo Copying ./$2 over $1/$STEEMD_CONFIG
+      cp ./$2 $1/$STEEMD_CONFIG
+      [ $? -ne 0 ] && echo FATAL: Failed to copy ./$2 over $1/$STEEMD_CONFIG file. && exit -1
+   else
+      echo WARNING: missing $STEEMD_CONFIG
+   fi
 }
 
-copy_config $TEST_WORK_PATH
-copy_config $REF_WORK_PATH
+function cleanup {
+   ARG=$1
+   if [ $REPLAY_PID1 ]
+   then
+     if [ $ARG -ne 0 ]; then
+        sleep 0.1 && kill -s SIGINT $REPLAY_PID1 &
+     fi
+     wait -n $REPLAY_PID1
+     [ $ARG -ne 0 -a $? -ne 0 ] && echo "FATAL: tested steemd failed to replay $BLOCK_LIMIT blocks." && EXIT_CODE=-1
+   fi
+   if [ $REPLAY_PID2 ]
+   then
+     if [ $ARG -ne 0 ]; then
+        sleep 0.1 && kill -s SIGINT $REPLAY_PID2 &
+     fi
+     wait -n $REPLAY_PID2
+     [ $ARG -ne 0 -a $? -ne 0 ] && echo "FATAL: reference steemd failed to replay $BLOCK_LIMIT blocks." && EXIT_CODE=-1
+   fi
+
+   if [ $ARG -ne 0 ]
+   then
+      EXIT_CODE=$ARG
+   fi
+}
+
+trap cleanup SIGINT SIGPIPE
+
+copy_config $TEST_WORK_PATH $STEEMD_CONFIG_TEST
+copy_config $REF_WORK_PATH  $STEEMD_CONFIG_REF
 
 echo Running "test instance" replay of $BLOCK_LIMIT blocks
 ( $TEST_STEEMD_PATH --replay --stop-replay-at-block $BLOCK_LIMIT -d $TEST_WORK_PATH ) & REPLAY_PID1=$!
@@ -36,12 +76,4 @@ echo Running "test instance" replay of $BLOCK_LIMIT blocks
 echo Running "reference instance" replay of $BLOCK_LIMIT blocks
 ( $REF_STEEMD_PATH --replay --stop-replay-at-block $BLOCK_LIMIT -d $REF_WORK_PATH ) & REPLAY_PID2=$!
 
-wait $REPLAY_PID1
-RET_VAL1=$?
-wait $REPLAY_PID2
-RET_VAL2=$?
-
-[ $RET_VAL1 -ne 0 ] && echo "FATAL: tested steemd failed to replay $BLOCK_LIMIT blocks." && exit -1
-[ $RET_VAL2 -ne 0 ] && echo "FATAL: reference steemd failed to replay $BLOCK_LIMIT blocks." && exit -1
-
-exit 0
+cleanup $EXIT_CODE
